services/kafka_consumer/requirements.txt
aiofiles
asgiref
click
fastapi==0.95.0   # Это стабильная версия, которая хорошо работает с зависимостями
h11
Jinja2
MarkupSafe
pydantic>=1.8,<2.0
starlette==0.26.1  # Совместимая версия с FastAPI 0.95.0
typing-extensions
uvicorn
kafka-python-ng
psycopg2